Organize public assets into js and css folders
