+++ public/js/dataviewer/tabManager.js
