+++ public/js/layout/resize.js
